api/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
APScheduler>=3.10.4
python-dotenv>=1.0.0
supabase>=2.9.0
httpx>=0.27.0
pika>=1.3.2
psutil>=5.9.0
pytz>=2024.1
croniter>=1.3.0
